Move files-related functions from utils/__init__.py to utils/files.py
